Move SmokeTester to separate file
